# GoExperts_Phase_3/APIs/docs/swagger.yaml
basePath: /
definitions:
  dto.CreateProductInput:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
...
  dto.UpdateProductInput:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
...
  entity.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
  description: API de estudos (Users e Products) usando net/http, GORM e JWT.
  title: Go Experts API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Lista produtos com paginação e ordenação via query string.
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: limit
        type: integer
      - description: Ordenação (asc|desc)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Lista produtos (paginado)
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Cria um produto (rota protegida).
      parameters:
      - description: Dados do produto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cria um produto
      tags:
      - Products
  /products/{id}:
    delete:
      description: Remove um produto pelo ID (rota protegida).
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove um produto
      tags:
      - Products
    get:
      description: Retorna um produto pelo ID (rota protegida).
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Busca produto por ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Atualiza parcialmente um produto pelo ID (rota protegida).
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Atualiza um produto
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {seu_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"